feat: 졸업학점 계산기 적용 학생 정보 수정 API 추가 (#1275)
* feat: 기존 학생 수정 API 졸학계 관련 로직 수정

* fix: 미사용 메소드 삭제

* feat: API 명세서 작성

* feat: 졸업학점 계산기 전용 학생 정보 수정 API 구현

* test: 테스트 수정

* chore: 스웨거 설명 수정

* chore: 스웨거 설명 수정

* refactor: 내부 로직 수정

* docs: 주석 작성

* feat: 리뷰 반영 및 로직 추가

* docs: 주석 추가

* style: 코드 포멧팅

* feat: flyway 추가

* chore: 초기값 설정 로직 삭제

* chore 리뷰 반영

* chore: flyway 번호 수정

+++ src/main/java/in/koreatech/koin/domain/graduation/service/GraduationService.java
@@ -103,10 +103,10 @@ public class GraduationService {
     @Transactional
     public void resetStudentCourseCalculation(Student student, Major newMajor) {
         // 기존 학생 졸업요건 계산 정보 삭제
-        studentCourseCalculationRepository.findByUserId(student.getUser().getId())
-            .ifPresent(studentCourseCalculation -> {
-                studentCourseCalculationRepository.deleteAllByUserId(student.getUser().getId());
-            });
+        List<StudentCourseCalculation> studentCourseCalculations = studentCourseCalculationRepository.findAllByUserId(student.getUser().getId());
+        if (!studentCourseCalculations.isEmpty()) {
+            studentCourseCalculationRepository.deleteAllByUserId(student.getUser().getId());
+        }
 
         initializeStudentCourseCalculation(student, newMajor);
